test: 테스트 오류 수정

+++ src/test/kotlin/model/BettingManagerTest.kt
@@ -43,6 +43,6 @@ class BettingManagerTest {
         val bettingAmount = 10000
         bettingManager.placeBet(player, bettingAmount)
 
-        assertThat(bettingManager.getProfit(player)).isEqualTo(bettingAmount)
+        assertThat(bettingManager.getProfit(player).toInt()).isEqualTo(bettingAmount)
     }
 }
